clean up & rename provider
